+++ src/GiapTech.LangCenter.Infrastructure/Persistence/Migrations/20260920062220_MotPhienMoiTaiKhoan.cs
@@ -1,0 +1,29 @@
+﻿using System;
+using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace GiapTech.LangCenter.Infrastructure.Persistence.Migrations
+{
+    /// <inheritdoc />
+    public partial class MotPhienMoiTaiKhoan : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<Guid>(
+                name: "phien_hien_tai",
+                table: "TAI_KHOAN",
+                type: "uuid",
+                nullable: true);
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "phien_hien_tai",
+                table: "TAI_KHOAN");
+        }
+    }
+}